services/backend/requirements.txt
# Reason: Define Python dependencies for FastAPI backend
# These packages provide web framework, JWT auth, WebSocket, and database connectivity

# Web framework and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# JWT authentication with RS256 support
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Password hashing
python-multipart==0.0.6

# WebSocket support (included in uvicorn[standard])
websockets==12.0

# Database connectivity
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Kafka client for consuming transactions
kafka-python==2.0.2

# Redis client and JSON Schema validation for pre-ingest checks
redis==4.6.0
jsonschema==4.19.0

# Vault client
hvac==1.1.1

# Utilities
python-dotenv==1.0.0
